refactor(onboard): move sandbox management primitives
